Move sequential_taskset.py into user module
