+++ locust/__init__.py
@@ -1,7 +1,7 @@
+from .user.sequential_taskset import SequentialTaskSet
 from .user.task import task, TaskSet
 from .user.users import HttpUser, User
 from .event import Events
-from .sequential_taskset import SequentialTaskSet
 from .wait_time import between, constant, constant_pacing
 
 events = Events()
